Fixed problems with chunks

+++ src/main/java/com/molina/cvmfs/common/PathHash.java
@@ -5,19 +5,19 @@ package com.molina.cvmfs.common;
  */
 public class PathHash {
 
-    private int hash1;
-    private int hash2;
+    private long hash1;
+    private long hash2;
 
-    public PathHash(int hash1, int hash2) {
+    public PathHash(long hash1, long hash2) {
         this.hash1 = hash1;
         this.hash2 = hash2;
     }
 
-    public int getHash1() {
+    public long getHash1() {
         return hash1;
     }
 
-    public int getHash2() {
+    public long getHash2() {
         return hash2;
     }
 }
